Adding the title after all other elements
Adding the title at the start was causing remeasurements of the actors
which, after adding action lines and other stuff, was producing
producing unnecessary space.  Now, the actors take up as much space as
needed, and if more space is required for the title, that space is taken
up as left and right padding.